remove most pragmas that was disabling warnings fix some warnings

+++ MyGUIEngine/include/MyGUI_RotatingSkin.h
@@ -77,7 +77,7 @@ namespace MyGUI
 
 		int _cropRotatedRectangle(FloatPoint* _baseVerticiesPos);
 		enum Side {Left, Right, Top, Bottom};
-		void _cropRotatedRectangleSide(std::vector<FloatPoint>& _verticies, float _sideCoord, Side _side);
+		void _cropRotatedRectangleSide(std::vector<FloatPoint>& _verticies, int _sideCoord, Side _side);
 
 		// get point position relative to rectangle
 		FloatPoint _getPositionInsideRect(const FloatPoint& _point, const FloatPoint& _corner0, const FloatPoint& _corner1, const FloatPoint& _corner2);
